+++ ts/table-grid/table-grid.scope.ts
@@ -1,0 +1,26 @@
+/// <reference path="../references.ts" />
+
+module Lui.Directives {
+
+	"use strict";
+
+	export interface IDataGridScope extends angular.IScope {
+		canvasHeight: any;
+		datas: any[];
+		filteredAndOrderedRows: any[];
+		fixedHeaderRows: TableGrid.Header[][];
+		fixedRowDefinition: TableGrid.Header[];
+		header: TableGrid.Tree;
+		leftFilters: {header: TableGrid.Header, value: string}[];
+		rightFilters: {header: TableGrid.Header, value: string}[];
+		scrollableHeaderRows: TableGrid.Header[][];
+		scrollableRowDefinition: TableGrid.Header[];
+		selected: { orderBy: TableGrid.Header, reverse: boolean };
+		visibleRows: any[];
+
+		updateFilterBy(header: TableGrid.Header, index: number): void;
+		updateOrderBy(header: TableGrid.Header): void;
+		updateVirtualScroll(): void;
+	}
+
+}